+++ Sources/KeyboardKit/Settings/KeyboardSettingsUrlProvider.swift
@@ -1,0 +1,36 @@
+//
+//  KeyboardSettingsUrlProvider.swift
+//  KeyboardKit
+//
+//  Created by Daniel Saidi on 2020-03-19.
+//  Copyright © 2020 Daniel Saidi. All rights reserved.
+//
+
+#if os(iOS) || os(tvOS)
+import Foundation
+import UIKit
+
+/**
+ This protocol can be implemented by any type that should be
+ able to resolve a URL to an app's keyboard settings page in
+ System Settings.
+
+ This protocol is implemented by `URL`.
+ */
+public protocol KeyboardSettingsUrlProvider {}
+
+public extension KeyboardSettingsUrlProvider {
+
+    /**
+     The url to the app's settings screen in System Settings.
+
+     If the app has no custom settings screen, this url will
+     open the main system settings screen.
+     */
+    static var keyboardSettings: URL? {
+        URL(string: UIApplication.openSettingsURLString)
+    }
+}
+
+extension URL: KeyboardSettingsUrlProvider {}
+#endif